Moving installs to higher folder
